float -> double

+++ src/br/ufrj/dcc/so/cliente/opcoes/OpcaoDOC.java
@@ -8,7 +8,7 @@ public class OpcaoDOC extends Opcao {
 	private String banco;
 	private String agencia;
 	private String conta;
-	private float valor;
+	private double valor;
 	
 	@Override
 	public void receberParametros() {
@@ -22,7 +22,7 @@ public class OpcaoDOC extends Opcao {
 		conta = in.nextLine();
 		
 		CLI.console("Digite o valor:");
-		valor = in.nextFloat();
+		valor = in.nextDouble();
 	}
 
 	@Override
